Pragma Engine Release 3.0
- Deleted ZombieGame
- Upgraded BallGame (colors & performance)
- Added NinjaPlaformer with platform input
- OPTIMALIED ENGINE

+++ BallGame/Grid.cpp
@@ -7,13 +7,13 @@ Grid::Grid(int width, int height, int cellSize) :
     m_width(width),
     m_height(height),
     m_cellSize(cellSize) {
-    m_numXCells = ceil((float)m_width / m_cellSize);
-    m_numYCells = ceil((float)m_height / m_cellSize);
+    m_numXCells = (int)ceil((float)m_width / m_cellSize);
+    m_numYCells = (int)ceil((float)m_height / m_cellSize);
 
     // Allocate all the cells
     const int BALLS_TO_RESERVE = 20;
     m_cells.resize(m_numYCells * m_numXCells);
-    for (int i = 0; i < m_cells.size(); i++) {
+    for (size_t i = 0; i < m_cells.size(); i++) {
         m_cells[i].balls.reserve(BALLS_TO_RESERVE);
     }
 }
@@ -56,7 +56,7 @@ void Grid::removeBallFromCell(Ball* ball) {
     balls[ball->cellVectorIndex] = balls.back();
     balls.pop_back();
     // Update vector index
-    if (ball->cellVectorIndex < balls.size()) {
+    if ((size_t)ball->cellVectorIndex < balls.size()) {
         balls[ball->cellVectorIndex]->cellVectorIndex = ball->cellVectorIndex;
     }
     // Set the index of ball to -1
